+++ packages/upload-media/src/store/constants.ts
@@ -27,6 +27,7 @@ export const CLIENT_SIDE_SUPPORTED_MIME_TYPES: readonly string[] = [
 	'image/gif',
 	'image/webp',
 	'image/avif',
+	'image/jxl',
 ] as const;
 
 /**