+++ packages/upload-media/src/store/types.ts
@@ -337,4 +337,4 @@ export interface SideloadAdditionalData extends AdditionalData {
 	image_size: string | string[];
 }
 
-export type ImageFormat = 'jpeg' | 'webp' | 'avif' | 'png' | 'gif';
+export type ImageFormat = 'jpeg' | 'webp' | 'avif' | 'png' | 'gif' | 'jxl';