+++ packages/vips/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### New Features
+
+-   Bundle `vips-jxl.wasm` to enable client-side JPEG XL (JXL) encoding and decoding.
+
 ## 1.4.0 (2026-04-15)
 
 ## 1.3.0 (2026-04-01)